% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dphixy.R
\name{dphixy}
\alias{dphixy}
\title{Scaled Bivariate Normal Density}
\usage{
dphixy(x, y, rho)
}
\arguments{
\item{x, y}{points value.}

\item{rho}{correlation coefficient.}
}
\value{
the density value.
}
\description{
Bivariate normal density with mean 0 variance 1.
}
\examples{
library(mvtnorm)
dmvnorm(c(1,-1),sigma = matrix(c(1,0.5,0.5,1),2,2))
dphixy(1,-1,0.5)
}
