% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esti_thre.R
\name{est_thre}
\alias{est_thre}
\title{Thresholds Estimation}
\usage{
est_thre(X)
}
\arguments{
\item{X}{a ordinal series.}
}
\value{
the estimated value for thresholds.
}
\description{
Function to calculate thresholds from ordinal variables.
}
\examples{
library(mvtnorm)
set.seed(1997)
R1 = gen_CCM(4)
n = 1000
indc = 3:4
thresholds = list(c(),c(),c(-1),c(1))
data1 = gen_mixed(n,R1,indc,thresholds=thresholds)$observed
est_thre(data1[,3])
est_thre(data1[,4])
}
