% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MCCM_est.R
\name{summary_MCCM_est}
\alias{summary_MCCM_est}
\title{Summary a MCCM Estimation Result}
\usage{
summary_MCCM_est(out_MCCM)
}
\arguments{
\item{out_MCCM}{output of function MCCM_est.}
}
\value{
The summary of estimation.
}
\description{
Display the estimated correlation matrix and std matrix for a MCCM_est list.
}
\seealso{
\link{MCCM_est},
\link{draw_correlation_matrix}
}
